switch ctd and tchain
